pyproject.toml
[project]
name = "marcus"
version = "0.1.0"
description = "Marcus — Real-Time Stoic Voice Agent. An AI voice companion that responds as Marcus Aurelius with deep baritone voice, finetuned on primary Stoic texts, with RL feedback loop."
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "mlx>=0.22",
    "mlx-lm>=0.22",
    "mlx-audio>=0.2",
    "sounddevice>=0.5",
    "numpy>=1.26",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "typer>=0.12",
    "httpx>=0.27",
]

[project.optional-dependencies]
ui = [
    "gradio>=4.0",
    "fastrtc>=0.0.20",
]
train = [
    "datasets>=3.0",
    "anthropic>=0.40",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[project.scripts]
marcus = "marcus.ui.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/marcus"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.12",
]
